+++ nattrs/nested_attributes.py
@@ -1,5 +1,13 @@
+"""
+Functionality for performing recursive get/set/has/mutate/update
+on nested classes and dictionaries (interchangeably).
+
+"""
+
+from dataclasses import dataclass
+import re
 from functools import partial
-from typing import Callable, Union, Any
+from typing import Callable, Dict, List, Optional, Tuple, Union, Any
 from collections.abc import Mapping, MutableMapping
 
 # TODO Add tests
@@ -10,6 +18,13 @@ from collections.abc import Mapping, MutableMapping
 # TODO Allow setting custom object instead of dict? Would require some thinking but could be done
 # TODO Make nested_delattr
 # TODO For getattr -> allow different default for leaf
+# TODO: nested_update
+# TODO: wildcard (*) or regex in path leading to bulk op
+# TODO: For multiple regex matches, return a dict with path -> value?
+
+DOT_PLACEHOLDER = "[[DOT]]"  # Placeholder for dots inside regex terms
+
+#### Getter ####
 
 
 def nested_getattr(
@@ -17,11 +32,14 @@ def nested_getattr(
     attr: str,
     default: Any = None,
     allow_none: bool = False,
-):
+    regex: bool = False,
+) -> Optional[Union[Any, Dict[str, Any]]]:
     """
     Get object attributes/dict members recursively, given by dot-separated names.
 
     Pass `attr='x.a.o'` to get attribute "o" of attribute "a" of attribute "x".
+
+    Use regular expression to get all matching attribute paths.
 
     Parameters
     ----------
@@ -36,17 +54,34 @@ def nested_getattr(
         and so on.
     default : Any
         Value to return when one or more of the attributes were not found.
+        NOTE: ignored when `regex=True` in which case an empty dict is the default.
     allow_none : bool
         Whether to allow `obj` to be `None` (in the first call - non-recursively).
         When allowed, such a call would return `None`.
+    regex : bool
+        Whether to interpret attribute/member names wrapped in `{}`
+        as regex patterns. When multiple matches exist, all are returned
+        in a list. Note: The entire attribute name must be included
+        in the wrapper, otherwise the name is considered a "fixed" (non-regex)
+        name. This means, "{" and "}" can be used within the regex.
+        Dots within "{}" are respected (i.e. not considered path splits).
 
     Returns
     -------
-    Any
-        One of:
-            - The value of the final attribute/dict member.
-            - The default value
-            - `None`, when `obj` is `None` and `allow_none` is `True`.
+    Any or dict[str, Any] or `None`
+        When `obj` is `None` and `allow_none` is `True`: `None`
+
+        When `regex=False`: Any
+            One/more of:
+                - The value of the final attribute/dict member.
+                - The default value
+
+        When `regex=True`: Dict[str, Any]
+            Dict mapping matching attribute paths to the value of
+            their final attribute/dict members.
+            When no matches were found, an empty dict is returned.
+
+
 
     Examples
     --------
@@ -72,17 +107,43 @@ def nested_getattr(
 
     >>> nested_getattr(a, "b.o.p", default="not found")
     "not found"
+
+    Using regex to find attribute names / dict keys:
+
+    >>> nested_getattr(a, "b.{.*}.{.*}", regex=True)
+    {"b.c.d": 1}
+
     """
-    if not allow_none and obj is None:
-        raise TypeError("`obj` was `None`.")
+    if obj is None:
+        if not allow_none:
+            raise TypeError("`obj` was `None`.")
+        return None
+
+    # When regex-use is enabled
+    if regex:
+        # Extract terms for regex compilation
+        terms = [
+            a.replace(DOT_PLACEHOLDER, ".")
+            for a in _replace_dots_in_regex(attr).split(".")
+        ]
+        attr, regexes = _precompile_regexes(terms)
+        matches = _regex_nested_getattrs(
+            objs=[obj],
+            attr=attr,
+            regexes=regexes,
+            default=None,
+        )
+        return {
+            match.attr_name: match.value
+            for match in _flatten_matches(matches)
+            if isinstance(match, MatchResult)
+        }
+
+    # Separate method for non-regex version (faster, simpler)
     return _nested_getattr(obj=obj, attr=attr, default=default)
 
 
 def _nested_getattr(obj: Union[object, Mapping], attr: str, default: Any = None):
-    """
-    Inspired by:
-    https://programanddesign.com/python-2/recursive-getsethas-attr/
-    """
     if obj is None:
         return obj
     getter = _dict_getter if isinstance(obj, Mapping) else getattr
@@ -90,7 +151,109 @@ def _nested_getattr(obj: Union[object, Mapping], attr: str, default: Any = None)
         left, right = attr.split(".", 1)
     except:  # noqa: E722
         return getter(obj, attr, default)
-    return _nested_getattr(getter(obj, left, default), right, default)
+    return _nested_getattr(
+        obj=getter(obj, left, default),
+        attr=right,
+        default=default,
+    )
+
+
+def _regex_nested_getattrs(
+    objs: List[Union[object, Mapping]],
+    attr: str,
+    regexes: Dict[str, re.Pattern],
+    default: Any = None,
+):
+    return [
+        _regex_nested_getattr(
+            obj=obj,
+            attr=attr,
+            regexes=regexes,
+            default=default,
+        )
+        for obj in objs
+    ]
+
+
+def _regex_nested_getattr(
+    obj: Union[object, Mapping],
+    attr: str,
+    regexes: Dict[str, re.Pattern],
+    default: Any = None,
+):
+    prev_attr = ""
+    if isinstance(obj, MatchResult):
+        prev_attr = obj.attr_name
+        obj = obj.value
+    if obj is None:
+        return [None]
+
+    getter = _dict_getter if isinstance(obj, Mapping) else getattr
+    try:
+        left, right = attr.split(".", 1)
+    except:  # noqa: E722
+        return _match_and_get_attr(
+            obj=obj,
+            attr=attr,
+            getter=getter,
+            regexes=regexes,
+            default=default,
+            prev_attr=prev_attr,
+        )
+
+    return _regex_nested_getattrs(
+        _match_and_get_attr(
+            obj=obj,
+            attr=left,
+            getter=getter,
+            regexes=regexes,
+            default=default,
+            prev_attr=prev_attr,
+        ),
+        attr=right,
+        regexes=regexes,
+        default=default,
+    )
+
+
+def _match_and_get_attr(
+    obj: Any,
+    attr: str,
+    getter: Callable,
+    regexes: Dict[str, re.Pattern],
+    default: Any,
+    prev_attr: str = "",
+) -> List[Any]:
+    matches = []
+
+    if attr.startswith("{") and attr.endswith("}"):
+        regex_id = attr[1:-1]  # Extract the regex pattern
+        pattern = regexes[regex_id]
+        keys = obj.keys() if isinstance(obj, Mapping) else _get_class_attributes(obj)
+
+        for key in keys:
+            if re.fullmatch(pattern, key):
+                # Collect all matches
+                matches.append(
+                    MatchResult(
+                        attr_name=f"{prev_attr}.{key}" if prev_attr else key,
+                        value=getter(obj, key, default),
+                    )
+                )
+        return (
+            matches if matches else [default]
+        )  # Return all matches, or default if no match
+
+    # Fallback to non-regex case
+    return [
+        MatchResult(
+            attr_name=f"{prev_attr}.{attr}" if prev_attr else attr,
+            value=getter(obj, attr, default),
+        )
+    ]
+
+
+#### Has checker ####
 
 
 def nested_hasattr(obj: Union[object, Mapping], attr: str, allow_none: bool = False):
@@ -150,10 +313,6 @@ def nested_hasattr(obj: Union[object, Mapping], attr: str, allow_none: bool = Fa
 
 
 def _nested_hasattr(obj: Union[object, Mapping], attr: str):
-    """
-    Inspired by:
-    https://programanddesign.com/python-2/recursive-getsethas-attr/
-    """
     if obj is None:
         return False
     getter = _dict_getter if isinstance(obj, Mapping) else getattr
@@ -163,6 +322,9 @@ def _nested_hasattr(obj: Union[object, Mapping], attr: str):
     except:  # noqa: E722
         return has_checker(obj, attr)
     return _nested_hasattr(getter(obj, left, None), right)
+
+
+#### Setter ####
 
 
 def nested_setattr(
@@ -177,6 +339,9 @@ def nested_setattr(
     Pass `attr='x.a.o'` to set attribute "o" of attribute "a" of attribute "x".
 
     When `make_missing` is disabled, it requires all but the last attribute/member to already exist.
+
+    If you want type-checking of existing values before assignment (or similar checks),
+    consider using `nested_mutattr()` instead and include such checks in the mutator function.
 
     Parameters
     ----------
@@ -240,16 +405,26 @@ def nested_setattr(
     )
 
 
+#### Mutator ####
+
+
 def nested_mutattr(
     obj: Union[object, MutableMapping],
     attr: str,
     fn: Callable,
     is_inplace_fn: bool = False,
+    getter_default: Any = None,
 ) -> None:
     """
     Mutate object attribute/dict member by recursive lookup, given by dot-separated names.
 
     Pass `attr='x.a.o'` to mutate attribute "o" of attribute "a" of attribute "x".
+
+    When the attribute doesn't exist, the mutating `fn` will receive `None`
+    by default (see `getter_default`).
+
+    Tip: The mutating function can perform checks of typing (e.g. consistency between
+    the new and original values) or similar.
 
     Parameters
     ----------
@@ -272,7 +447,8 @@ def nested_mutattr(
         function to have made the relevant change when applied.
         E.g. useful for `numpy.ndarrays` that we do not
         wish to copy.
-
+    getter_default : Any
+        Value to pass to `fn()` when the attribute was not found.
 
     Examples
     --------
@@ -298,7 +474,11 @@ def nested_mutattr(
     >>> nested_getattr(a, "b.c.d")
     10
     """
-    old_val = nested_getattr(obj=obj, attr=attr)
+    old_val = nested_getattr(
+        obj=obj,
+        attr=attr,
+        default=getter_default,
+    )
     try:
         if is_inplace_fn:
             fn(old_val)
@@ -338,3 +518,79 @@ def _dict_setter(obj: MutableMapping, key: Any, val: Any):
 
 def _dict_has(obj: Mapping, key: Any):
     return key in obj
+
+
+def _replace_dots_in_regex(attr: str) -> str:
+    """Replace dots inside regex terms (inside `{}`) with a placeholder."""
+    result = []
+    inside_regex = False
+
+    for char in attr:
+        if char == "{":
+            inside_regex = True
+        elif char == "}":
+            inside_regex = False
+        if char == "." and inside_regex:
+            result.append(DOT_PLACEHOLDER)
+        else:
+            result.append(char)
+
+    return "".join(result)
+
+
+def _precompile_regexes(terms: List[str]) -> Tuple[str, Dict[str, re.Pattern]]:
+    """
+    Extract regex patterns wrapped in `{}` that cover entire attribute terms,
+    compile them, and replace them with placeholders like `regex_0`, `regex_1`, etc.
+
+    Parameters
+    ----------
+    terms : List[str]
+        List of the attribute names/keys.
+
+    Returns
+    -------
+    Tuple[str, List[re.Pattern]]
+        - The modified dot-joined attribute string with regex placeholders.
+        - List of compiled regexes.
+    """
+    regexes = {}
+    new_terms = []
+
+    for i, term in enumerate(terms):
+        # Check if the term is completely wrapped in {}
+        if term.startswith("{") and term.endswith("}"):
+            # Compile the regex inside the curly braces
+            pattern = term[1:-1]
+            compiled_regex = re.compile(pattern)
+            regexes[f"regex_{i}"] = compiled_regex
+            # Replace the regex term with a placeholder like 'regex_0'
+            new_terms.append("{" + f"regex_{i}" + "}")
+        else:
+            new_terms.append(term)
+
+    # Join the terms back into a dot-separated path
+    modified_attr = ".".join(new_terms)
+
+    return modified_attr, regexes
+
+
+@dataclass
+class MatchResult:
+    attr_name: str
+    value: Any
+
+
+def _get_class_attributes(obj):
+    """Return only user-defined attributes, excluding methods and built-in attributes."""
+    if hasattr(obj, "__dict__"):
+        return obj.__dict__.keys()
+    return []
+
+
+def _flatten_matches(lst):
+    for item in lst:
+        if isinstance(item, list):
+            yield from _flatten_matches(item)
+        else:
+            yield item